# server, tracing: Make traces redactable and default to "off"