--- a/pkg/kv/txn_test.go
+++ b/pkg/kv/txn_test.go
@@ -68,7 +68,6 @@ func TestTxnVerboseTrace(t *testing.T) {
 		`(?s)`+
 			`.*event:[^:]*:\d+ inside txn\n`+
 			`.*event:[^:]*:\d+ client\.Txn did AutoCommit\. err: <nil>\n`+
-			`.*\n`+
 			`.*event:[^:]*:\d+ txn complete.*`,
 		dump)
 	if err != nil {