@dhartunian 2021-12-08
@andreimatei this addresses most of your concerns from #73117 and is a much smaller PR. I will update the one on `release-21.2` tomorrow and respond to the open threads.

You were correct that the use of `redact.Sprintf` vs `fmt.Fprintf` in `spanInner.Recordf` did not significantly affect the runtime. I ran a narrower benchmark (`BenchmarkEventf_WithVerboseTraceSpan`) and found that the largest performance gain by far is in using the `LegacyString` method on the `logEntry`, which is now renamed to `FastString`.